// web-chat/front001/src/types/KorisnikType.ts
export default class KorisnikType {
    created_at?: string;
    email?: string;
    family_name?: string;
    given_name?: string;
    locale?: string;
    name?: string;
    nickname?: string;
    picture?: string;
    updated_at?: string;
    user_id?: string;
    last_login?: string;
    last_ip?: string;
    logins_count?: string;
}